vous faites chier

+++ src/main/java/dev/services/MissionService.java
@@ -32,8 +32,10 @@ import dev.exceptions.InvalidDateTransportMissionException;
 import dev.exceptions.InvalidIdException;
 import dev.exceptions.InvalidIdMissionException;
 import dev.exceptions.NoPrimeException;
+import dev.exceptions.SupressionImpossibleErreur;
 import dev.repositories.CollegueRepo;
 import dev.repositories.MissionRepo;
+import dev.repositories.NoteDeFraisRepo;
 
 @Service
 public class MissionService {
@@ -43,6 +45,8 @@ public class MissionService {
 
 	@Autowired
 	private CollegueRepo collegueRepo;
+	@Autowired
+	private NoteDeFraisRepo noteDeFrais;
 
 	public MissionService() {
 		super();
@@ -252,10 +256,16 @@ public class MissionService {
 	 */
 	public void deleteMission(Mission mission) {
 		if (this.missionRepo.existsById(mission.getId())) {
-			this.missionRepo.delete(mission);
+			if (this.noteDeFrais.findNoteDeFraisByMission(mission) == null) {
+				this.missionRepo.delete(mission);
 
+			} else {
+				throw new SupressionImpossibleErreur();
+
+			}
 		} else {
 			throw new InvalidIdException();
+
 		}
 
 	}
